refactor: move common function to utils
